page-optimizer: icons seperate
