+++ includes/admin/page-optimizer/src/app/page-optimizer/components/icons/category-icons.tsx
@@ -1,0 +1,56 @@
+import React from 'react';
+
+interface SvgProps {
+    cls?: string;
+}
+
+const width = 34;
+const height = 34;
+
+export const CSSDelivery = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#7F54B3"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M9.5 7.00689H8.07C7.53 7.00689 7.25 7.00689 7.16 7.16689C7.02665 7.35185 6.96938 7.58094 7 7.80689C7.00646 8.36465 6.97975 8.9223 6.92 9.47689C6.88923 9.73554 6.81127 9.98636 6.69 10.2169C6.52022 10.4949 6.28285 10.7254 6 10.8869C6.28821 11.0502 6.52669 11.2887 6.69 11.5769C6.85676 11.9408 6.94208 12.3366 6.94 12.7369L7 14.2969C7.00825 14.3648 7.03221 14.4299 7.07 14.4869C7.10698 14.5458 7.15868 14.5941 7.22 14.6269C7.48081 14.7409 7.76757 14.7824 8.05 14.7469H9.5V7.00689ZM11.9 14.7469H13.31C13.89 14.7469 14.11 14.7469 14.22 14.5969C14.3518 14.4297 14.416 14.2192 14.4 14.0069C14.3984 13.4154 14.4318 12.8244 14.5 12.2369C14.5414 11.9829 14.6295 11.7387 14.76 11.5169C14.9196 11.2689 15.1399 11.0658 15.4 10.9269C15.1402 10.7799 14.9151 10.5786 14.74 10.3369C14.6214 10.1426 14.5369 9.92958 14.49 9.70689C14.4433 9.24022 14.4133 8.80689 14.4 8.40689C14.3969 8.05251 14.3736 7.6986 14.33 7.34689C14.2882 7.24858 14.2141 7.16746 14.12 7.11689C13.858 7.02075 13.5781 6.98321 13.3 7.00689H11.9V14.7469Z" fill="#7148A2" stroke="#E7E7E7"/>
+    </svg>
+);
+
+export const JavascriptDelivery = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#FDC20A"/>
+        <path d="M13.0614 8.4258L15.315 10.6924C15.4283 10.8063 15.4283 10.9909 15.315 11.1048L13.0614 13.3714M7.93851 13.3714L5.68491 11.1048C5.57166 10.9909 5.57166 10.8063 5.68491 10.6924L7.93851 8.4258M11.1662 6.70001L9.92659 15" stroke="white" strokeLinecap="round"/>
+    </svg>
+);
+
+export const ImageDeliverySVG = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#11BFE6"/>
+        <circle cx="8.40002" cy="12.6" r="3" fill="#0E9AB9" stroke="white"/>
+        <circle cx="13.6501" cy="7.35003" r="1.95" fill="#0E9AB9" stroke="white"/>
+    </svg>
+);
+
+export const FontDelivery = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#295ECF"/>
+        <path d="M11.1095 12.9899C11.053 13.2628 10.9683 13.4416 10.8366 13.5357C10.7049 13.6298 10.4696 13.6674 10.1309 13.6674L10.0933 14.8153C10.7895 14.8153 11.3165 14.693 11.674 14.4296C12.0222 14.1661 12.2574 13.7239 12.3891 13.084L13.1042 9.40501H13.904L14.1675 8.44526H13.2924L13.4053 7.89952C13.4618 7.57961 13.6688 7.41024 14.0263 7.41024C14.2522 7.41024 14.4498 7.44788 14.6003 7.50433L14.9955 6.48813C14.6944 6.36581 14.3463 6.29994 13.9511 6.29994C13.49 6.29994 13.0948 6.45049 12.7561 6.73277C12.4174 7.01505 12.2009 7.38201 12.1163 7.83366L11.9939 8.44526H11.2506L11.0624 9.40501H11.8058L11.1095 12.9899Z" fill="white"/>
+        <path d="M6.30005 12.9617H7.67381L8.16309 10.4494H10.2049L10.4213 9.31085H8.38891L8.64296 8.00296H10.9388L11.1647 6.84562H7.49503L6.30005 12.9617Z" fill="white"/>
+    </svg>
+);
+
+export const CloudDelivery = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#1EB47D"/>
+        <path d="M7.70002 14.7C6.9279 14.7 6.26812 14.4244 5.72068 13.8731C5.17324 13.3219 4.89969 12.6481 4.90002 11.8519C4.90002 11.1694 5.09942 10.5612 5.49821 10.0275C5.89699 9.49374 6.41881 9.15249 7.06366 9.00374C7.27578 8.19874 7.70002 7.54686 8.33639 7.04811C8.97275 6.54936 9.69396 6.29999 10.5 6.29999C11.4928 6.29999 12.335 6.65664 13.0266 7.36994C13.7183 8.08324 14.064 8.95159 14.0637 9.97499C14.6491 10.045 15.135 10.3054 15.5212 10.7562C15.9074 11.207 16.1004 11.7341 16.1 12.3375C16.1 12.9937 15.8772 13.5516 15.4316 14.0112C14.986 14.4707 14.4451 14.7003 13.8091 14.7H7.70002Z" fill="#157F58" stroke="white"/>
+        <path d="M10.08 9.72018V13.3H9.36371V11.2299L9.38174 10.5172C9.26837 10.6369 9.18591 10.7194 9.13438 10.7647L8.7453 11.0939L8.40002 10.6396L10.08 9.09996V9.72018Z" fill="white"/>
+        <path d="M10.92 12.6798L10.92 9.09999L11.6363 9.09999L11.6363 11.1701L11.6183 11.8828C11.7316 11.7631 11.8141 11.6806 11.8656 11.6352L12.2547 11.3061L12.6 11.7604L10.92 13.3L10.92 12.6798Z" fill="white"/>
+    </svg>
+);
+
+export const PageCache = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 21 21" fill="none">
+        <path d="M14.2056 0H6.79443C3.04197 0 0 3.04197 0 6.79443V14.2056C0 17.958 3.04197 21 6.79443 21H14.2056C17.958 21 21 17.958 21 14.2056V6.79443C21 3.04197 17.958 0 14.2056 0Z" fill="#FF7D00"/>
+        <path d="M16.1 8.89999H16.6V8.39999V8.29999C16.6 6.91928 15.4807 5.79999 14.1 5.79999H6.90002C5.51931 5.79999 4.40002 6.91928 4.40002 8.29999V8.39999V8.89999H4.90002H16.1Z" fill="#C76200" stroke="white"/>
+        <path d="M4.90002 8.59991H4.40002V9.09991V12.6999C4.40002 14.0806 5.51931 15.1999 6.90002 15.1999H14.1C15.4807 15.1999 16.6 14.0806 16.6 12.6999V9.09991V8.59991H16.1H4.90002Z" fill="#C76200" stroke="white"/>
+        <path d="M11.5 10L9.54585 11.3052L10.3377 11.7893L9.5 13L11.4542 11.7134L10.6633 11.2309L11.5 10Z" fill="white"/>
+    </svg>
+);
