+++ includes/admin/page-optimizer/src/app/page-optimizer/components/icons/gear-icons.tsx
@@ -1,0 +1,45 @@
+import React, {FC} from 'react';
+
+interface SvgProps {
+    cls?: string;
+}
+
+const width = 34;
+const height = 34;
+
+export const Starter:FC<SvgProps> = ({cls}) => (
+    <svg className={`${cls}`} width="84" height="75" viewBox="0 0 74 65" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path fillRule="evenodd" clipRule="evenodd" d="M32.7568 64.6651C49.946 63.7025 63.2062 49.2644 62.224 32.4198C62.0603 29.6926 61.5691 27.1258 60.7506 24.559L60.2595 24.8799C61.078 27.4467 61.5691 30.1739 61.5691 32.9011C61.5691 47.6601 49.2912 59.5315 34.2302 59.5315C33.7391 59.5315 33.248 59.5315 32.7568 59.5315V64.6651Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M5.41602 51.3504C10.982 59.3716 20.1496 64.3448 29.972 64.6656V59.2112C22.2778 57.9278 15.7295 54.5589 11.4731 47.9815L5.41602 51.3504Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M2.63385 20.3886C-1.2951 29.6932 -0.803979 40.1208 3.9435 48.9441L10.1643 45.4148C7.05392 38.8374 7.21763 31.2974 10.6555 24.8805L2.63385 20.3886Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M29.9723 0.335922C18.8403 1.61932 9.0179 8.19671 3.7793 17.8222L11.9646 22.4745C16.0573 16.5388 22.6055 12.849 29.8086 12.3677V0.335922H29.9723Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M60.9117 13.1694C54.1998 4.66694 43.7226 -0.306215 32.918 0.0146335V12.2069C38.8114 12.5277 44.2137 14.7737 48.4701 18.7843L60.9117 13.1694Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M31.4475 29.3722C34.5579 29.3722 37.1772 31.7785 37.1772 34.987C37.1772 38.0351 34.7216 40.6019 31.4475 40.6019C28.3371 40.6019 25.7178 38.1955 25.7178 34.987C25.7178 31.9389 28.3371 29.3722 31.4475 29.3722Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M67.4615 14.7733L50.436 26.1635L33.4106 37.3931L31.9372 34.8264L30.4639 32.2596L48.9627 23.5967L67.4615 14.7733Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M73.1907 11.7255L68.4433 16.5382H68.2795L68.1158 16.3778C67.9521 15.7361 67.7884 15.0944 67.461 14.6131C67.1336 14.1319 66.8062 13.6506 66.3151 13.1693C66.1514 13.1693 66.1514 13.0089 66.1514 12.8485C66.1514 12.8485 66.1514 12.8485 66.3151 12.688L73.027 11.0838C73.1907 11.0838 73.1907 11.0838 73.1907 11.2442C73.3544 11.5651 73.3544 11.7255 73.1907 11.7255Z" fill="#673B9B"/>
+    </svg>
+);
+export const Accelerate:FC<SvgProps> = ({cls}) => (
+    <svg className={`${cls}`} width="84" height="75" viewBox="0 0 74 65" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path fillRule="evenodd" clipRule="evenodd" d="M32.7568 64.6651C49.946 63.7025 63.2062 49.2644 62.224 32.4198C62.0603 29.6926 61.5691 27.1258 60.7506 24.559L60.2595 24.8799C61.078 27.4467 61.5691 30.1739 61.5691 32.9011C61.5691 47.6601 49.2912 59.5315 34.2302 59.5315C33.7391 59.5315 33.248 59.5315 32.7568 59.5315V64.6651Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M5.41602 51.3504C10.982 59.3716 20.1496 64.3448 29.972 64.6656V59.2112C22.2778 57.9278 15.7295 54.5589 11.4731 47.9815L5.41602 51.3504Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M2.63385 20.3886C-1.2951 29.6932 -0.803979 40.1208 3.9435 48.9441L10.1643 45.4148C7.05392 38.8374 7.21763 31.2974 10.6555 24.8805L2.63385 20.3886Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M29.9723 0.335922C18.8403 1.61932 9.0179 8.19671 3.7793 17.8222L11.9646 22.4745C16.0573 16.5388 22.6055 12.849 29.8086 12.3677V0.335922H29.9723Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M60.9117 13.1694C54.1998 4.66694 43.7226 -0.306215 32.918 0.0146335V12.2069C38.8114 12.5277 44.2137 14.7737 48.4701 18.7843L60.9117 13.1694Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M31.4475 29.3722C34.5579 29.3722 37.1772 31.7785 37.1772 34.987C37.1772 38.0351 34.7216 40.6019 31.4475 40.6019C28.3371 40.6019 25.7178 38.1955 25.7178 34.987C25.7178 31.9389 28.3371 29.3722 31.4475 29.3722Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M67.4615 14.7733L50.436 26.1635L33.4106 37.3931L31.9372 34.8264L30.4639 32.2596L48.9627 23.5967L67.4615 14.7733Z" fill="#673B9B"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M73.1907 11.7255L68.4433 16.5382H68.2795L68.1158 16.3778C67.9521 15.7361 67.7884 15.0944 67.461 14.6131C67.1336 14.1319 66.8062 13.6506 66.3151 13.1693C66.1514 13.1693 66.1514 13.0089 66.1514 12.8485C66.1514 12.8485 66.1514 12.8485 66.3151 12.688L73.027 11.0838C73.1907 11.0838 73.1907 11.0838 73.1907 11.2442C73.3544 11.5651 73.3544 11.7255 73.1907 11.7255Z" fill="#673B9B"/>
+    </svg>
+);
+export const TurboMax:FC<SvgProps> = ({cls}) => (
+    <svg className={`${cls}`} width="84" height="75" viewBox="0 0 74 65" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path fillRule="evenodd" clipRule="evenodd" d="M32.7568 64.6651C49.946 63.7025 63.2062 49.2644 62.224 32.4198C62.0603 29.6926 61.5691 27.1258 60.7506 24.559L60.2595 24.8799C61.078 27.4467 61.5691 30.1739 61.5691 32.9011C61.5691 47.6601 49.2912 59.5315 34.2302 59.5315C33.7391 59.5315 33.248 59.5315 32.7568 59.5315V64.6651Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M5.41602 51.3504C10.982 59.3716 20.1496 64.3448 29.972 64.6656V59.2112C22.2778 57.9278 15.7295 54.5589 11.4731 47.9815L5.41602 51.3504Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M2.63385 20.3886C-1.2951 29.6932 -0.803979 40.1208 3.9435 48.9441L10.1643 45.4148C7.05392 38.8374 7.21763 31.2974 10.6555 24.8805L2.63385 20.3886Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M29.9723 0.335922C18.8403 1.61932 9.0179 8.19671 3.7793 17.8222L11.9646 22.4745C16.0573 16.5388 22.6055 12.849 29.8086 12.3677V0.335922H29.9723Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M60.9117 13.1694C54.1998 4.66694 43.7226 -0.306215 32.918 0.0146335V12.2069C38.8114 12.5277 44.2137 14.7737 48.4701 18.7843L60.9117 13.1694Z" fill="#DDDFE4"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M31.4475 29.3722C34.5579 29.3722 37.1772 31.7785 37.1772 34.987C37.1772 38.0351 34.7216 40.6019 31.4475 40.6019C28.3371 40.6019 25.7178 38.1955 25.7178 34.987C25.7178 31.9389 28.3371 29.3722 31.4475 29.3722Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M67.4615 14.7733L50.436 26.1635L33.4106 37.3931L31.9372 34.8264L30.4639 32.2596L48.9627 23.5967L67.4615 14.7733Z" fill="#5B3786"/>
+        <path fillRule="evenodd" clipRule="evenodd" d="M73.1907 11.7255L68.4433 16.5382H68.2795L68.1158 16.3778C67.9521 15.7361 67.7884 15.0944 67.461 14.6131C67.1336 14.1319 66.8062 13.6506 66.3151 13.1693C66.1514 13.1693 66.1514 13.0089 66.1514 12.8485C66.1514 12.8485 66.1514 12.8485 66.3151 12.688L73.027 11.0838C73.1907 11.0838 73.1907 11.0838 73.1907 11.2442C73.3544 11.5651 73.3544 11.7255 73.1907 11.7255Z" fill="#5B3786"/>
+    </svg>
+);
